ISU-5 Rewrite TypedWorker
ISU-5: Remove CallBackWorker
