+++ test/TypedWorker.spec.ts
@@ -1,15 +1,15 @@
-import { TypedWorker } from '../src/index'
+import { createWorker, ITypedWorker } from '../src/index'
 
 const range = n => Array.from({ length: n }, (value, key) => key)
 
 describe("TypedWorker", function () {
 
   describe("basic message passing", () => {
-    let result;
+    let result: number;
 
     beforeEach(function (done) {
-      const numberWorker = new TypedWorker(
-        (input: number) => 1000,
+      const numberWorker: ITypedWorker<number, number> = createWorker(
+        (input, cb) => cb(1000),
         (output) => {
           result = output
           done();
@@ -27,6 +27,27 @@ describe("TypedWorker", function () {
     });
   })
 
+  describe("simple cb with object", () => {
+    let objectWorker: ITypedWorker<{ a: number }, { b: number }>;
+    let result = {}
+    beforeEach(done => {
+      objectWorker = createWorker(
+        (input: { a: number }, cb) => {
+          cb({ b: input.a });
+        },
+        (output: { b: number }) => {
+          result = output;
+          done();
+        }
+      )
+      objectWorker.postMessage({ a: 10 });
+    });
+
+    it("simple object return", function () {
+      expect(result).toEqual({ b: 10 })
+    });
+  });
+
   describe("multi-messages", () => {
     describe("small message count", () => {
 
@@ -35,8 +56,8 @@ describe("TypedWorker", function () {
       const numberRangeSmall = range(10)
 
       beforeEach(function (done) {
-        const numberWorker = new TypedWorker(
-          (input: number) => input,
+        const numberWorker: ITypedWorker<number, number> = createWorker(
+          (input, cb) => cb(input),
           (output) => {
             result += output
             msgCountDown--
@@ -60,8 +81,8 @@ describe("TypedWorker", function () {
       let msgCountDown = numberRangeLarge.length
 
       beforeEach(function (done) {
-        const numberWorker = new TypedWorker(
-          (input: number) => input,
+        const numberWorker: ITypedWorker<number, number> = createWorker(
+          (input, cb) => cb(input),
           (output) => {
             result += output
             msgCountDown--
@@ -78,13 +99,43 @@ describe("TypedWorker", function () {
         expect(result).toEqual(expected)
       });
     })
+    describe("correct order", () => {
+      let multiReponse: ITypedWorker<number, number>;
+      let result: number[] = [];
+      beforeEach(done => {
+        result = [];
+        multiReponse = createWorker(
+          (input: number, cb) => {
+            cb(input);
+          },
+          (output: number) => {
+            result.push(output);
+            if (output === 3) {
+              done();
+            }
+          }
+        )
+        multiReponse.postMessage(1);
+        multiReponse.postMessage(2);
+        multiReponse.postMessage(3);
+      });
+
+
+      it("correct length", function () {
+        expect(result.length).toEqual(3);
+      });
+
+      it("order is correct", function () {
+        expect(result).toEqual([1, 2, 3]);
+      });
+    })
   })
 
   describe("handles large input", () => {
     let result;
     beforeEach(function (done) {
-      const numberWorker = new TypedWorker(
-        (input: number[]) => input,
+      const numberWorker: ITypedWorker<number[], number[]> = createWorker(
+        (input: number[], cb) => cb(input),
         (output) => {
           result = output.reduce((c, p) => c + p)
           done();
@@ -101,8 +152,8 @@ describe("TypedWorker", function () {
 
   describe("Termination", () => {
     let msgCounter = 0;
-    let numberWorker: TypedWorker<number, number>;
-    numberWorker = new TypedWorker(
+    let numberWorker: ITypedWorker<number, number>;
+    numberWorker = createWorker(
       (input: number) => {
         msgCounter += 1
         return input;
@@ -113,6 +164,29 @@ describe("TypedWorker", function () {
       numberWorker.terminate();
       numberWorker.postMessage(1);
       expect(msgCounter).toEqual(0)
+    });
+  })
+
+  describe("Termination", () => {
+    let msgCounter = 0;
+    let numberWorker: ITypedWorker<number, number>;
+    beforeEach(done => {
+      numberWorker = createWorker(
+        (input: number, cb) => {
+          cb(input);
+        },
+        (output) => {
+          numberWorker.terminate()
+          msgCounter = output;
+          setTimeout(done, 500)
+        }
+      )
+      numberWorker.postMessage(1);
+      numberWorker.postMessage(2);
+    });
+
+    it("stops handling messages", function () {
+      expect(msgCounter).toEqual(1)
     });
   })
 });
